move routes test files out of component folder
