+++ src/client/app/test/__tests__/routes/home.test.tsx
@@ -1,12 +1,12 @@
 import { render, screen } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
-import { setupAxiosMock } from '../../../util/axiosMockUtils.ts';
-import URLConstants from '../../../../util/constants/url-constants.ts';
-import Home from '../../../../routes/home.tsx';
-import { CategoriesContext } from '../../../../store/categories-context.tsx';
-import { mockCategoriesContextValue } from '../../../__mocks__/store/categories-context-mock.ts';
-import { mockCartContextValue } from '../../../__mocks__/store/cart-context-mock.ts';
-import { CartContext } from '../../../../store/cart-context.tsx';
+import { setupAxiosMock } from '../../util/axiosMockUtils.ts';
+import URLConstants from '../../../util/constants/url-constants.ts';
+import Home from '../../../routes/home.tsx';
+import { CategoriesContext } from '../../../store/categories-context.tsx';
+import { mockCategoriesContextValue } from '../../__mocks__/store/categories-context-mock.ts';
+import { mockCartContextValue } from '../../__mocks__/store/cart-context-mock.ts';
+import { CartContext } from '../../../store/cart-context.tsx';
 
 let container: HTMLElement;
