+++ src/client/app/test/__tests__/routes/login.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, render, screen, waitFor } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
-import Login from '../../../../routes/login';
+import Login from '../../../routes/login';
 
 let container: HTMLElement;
