+++ src/client/app/test/__tests__/routes/manage/manage-actions.test.tsx
@@ -1,11 +1,11 @@
 import { fireEvent, render, screen, waitFor } from '@testing-library/react';
 import { MemoryRouter, Route, Routes } from 'react-router-dom';
-import { AuthContext } from '../../../../../store/auth-context';
-import Login from '../../../../../routes/login';
-import ManageActions from '../../../../../routes/manage/manage-actions';
-import ManageCategories from '../../../../../routes/manage/manage-categories';
-import { ACTIONS } from '../../../../../util/constants/manage-actions-types';
-import { type ActionsProps } from '../../../../../types/routes/manage';
+import { AuthContext } from '../../../../store/auth-context';
+import Login from '../../../../routes/login';
+import ManageActions from '../../../../routes/manage/manage-actions';
+import ManageCategories from '../../../../routes/manage/manage-categories';
+import { ACTIONS } from '../../../../util/constants/manage-actions-types';
+import { type ActionsProps } from '../../../../types/routes/manage';
 
 describe('ManageActions', () => {
   describe('when unauthenticated', () => {
