+++ src/client/app/test/__tests__/routes/shop/product.test.tsx
@@ -1,13 +1,13 @@
 import { fireEvent, render, screen, waitFor } from '@testing-library/react';
 import { MemoryRouter, Routes, Route, Router } from 'react-router-dom';
 import { createMemoryHistory } from 'history';
-import { setupAxiosMock } from '../../../../util/axiosMockUtils';
-import URLConstants from '../../../../../util/constants/url-constants';
-import Shop from '../../../../../routes/shop/shop';
-import Product from '../../../../../routes/shop/product';
-import Error from '../../../../../routes/error';
-import { ProductsContext } from '../../../../../store/products-context';
-import { mockProductsContextValue } from '../../../../__mocks__/store/products-context-mock';
+import { setupAxiosMock } from '../../../util/axiosMockUtils';
+import URLConstants from '../../../../util/constants/url-constants';
+import Shop from '../../../../routes/shop/shop';
+import Product from '../../../../routes/shop/product';
+import Error from '../../../../routes/error';
+import { ProductsContext } from '../../../../store/products-context';
+import { mockProductsContextValue } from '../../../__mocks__/store/products-context-mock';
 
 const renderContentWithValidId = () => {
   return render(
