+++ src/client/app/test/__tests__/routes/shop/shop.test.tsx
@@ -1,15 +1,15 @@
 import { render, screen } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
-import { setupAxiosMock } from '../../../../util/axiosMockUtils';
-import URLConstants from '../../../../../util/constants/url-constants';
-import Shop from '../../../../../routes/shop/shop';
-import { CategoriesContext } from '../../../../../store/categories-context';
-import { ProductsContext } from '../../../../../store/products-context';
-import { mockCategoriesContextValue } from '../../../../__mocks__/store/categories-context-mock';
-import { mockProductsContextValue } from '../../../../__mocks__/store/products-context-mock';
-import { renderShopContextProps } from '../../../../__mocks__/store/context-mock';
+import { setupAxiosMock } from '../../../util/axiosMockUtils';
+import URLConstants from '../../../../util/constants/url-constants';
+import Shop from '../../../../routes/shop/shop';
+import { CategoriesContext } from '../../../../store/categories-context';
+import { ProductsContext } from '../../../../store/products-context';
+import { mockCategoriesContextValue } from '../../../__mocks__/store/categories-context-mock';
+import { mockProductsContextValue } from '../../../__mocks__/store/products-context-mock';
+import { renderShopContextProps } from '../../../__mocks__/store/context-mock';
 
-jest.mock('../../../../../components/card/card-product/helpers');
+jest.mock('../../../../components/card/card-product/helpers');
 
 const renderContext = (props: renderShopContextProps) => {
   const { categoriesContextValue, productsContextValue } = props;
